LayoutTests/html5lib/resources/adoption02.dat

#data
<b>1<i>2<p>3</b>4
#errors
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <i>
|         "2"
|     <i>
|       <p>
|         <b>
|           "3"
|         "4"

#data
<b id="b">1<i id="i">2<p>3<script>document.getElementById("b").id = "c";document.getElementById("i").id = "j"</script></b>4
#errors
#document
| <html>
|   <head>
|   <body>
|     <b>
|       id="c"
|       "1"
|       <i>
|         id="j"
|         "2"
|     <i>
|       id="i"
|       <p>
|         <b>
|           id="b"
|           "3"
|           <script>
|             "document.getElementById("b").id = "c";document.getElementById("i").id = "j""
|         "4"

#data
<a><div><style></style><address><a>
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|         <style>
|       <address>
|         <a>
|         <a>
